+++ VersionHistory.md
@@ -11,6 +11,8 @@
 * Improve exit code exception.
 * Clean all `tools` projects but keep `tools/bin`.
 * Support `--verbosity` command-line option.
+* Add target `format` if `dotnet-format` is installed locally.
+* Add targets `cleanup` and `inspect` if `JetBrains.ReSharper.GlobalTools` is installed locally.
 
 ### 4.8.0